requirements.txt
ccxt>=4.2.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
flask>=3.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
flask-cors>=4.0.0
aiohttp>=3.9.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
matplotlib>=3.7.0
